+++ packages/gensx-client/rollup.config.js
@@ -2,8 +2,4 @@ import { createConfigs } from "../../create-rollup-config.js";
 
 const external = ["@gensx/core"];
 
-export default createConfigs(
-  ["src/index.ts"],
-  "@gensx/client",
-  external,
-);
+export default createConfigs(["src/index.ts"], "@gensx/client", external);