+++ packages/gensx-react/rollup.config.js
@@ -2,8 +2,4 @@ import { createConfigs } from "../../create-rollup-config.js";
 
 const external = ["@gensx/client", "react", "react-dom"];
 
-export default createConfigs(
-  ["src/index.ts"],
-  "@gensx/react",
-  external,
-);
+export default createConfigs(["src/index.ts"], "@gensx/react", external);